feat: Buscar Professores

+++ app/src/main/java/com/example/hp_api/model/VarinhaModel.kt
@@ -3,5 +3,5 @@ package com.example.hp_api.model
 data class VarinhaModel(
     val wood: String,
     val core: String,
-    val length: Int
+    val length: Double
 )
